Implement POSIX controlling terminals and /dev/tty
A handle to a terminal is passed via a private interface from the terminal
service to the POSIX service when the terminal is set as controlling terminal
for a session. A process can then request to open a handle to its controlling
terminal from the POSIX service.

For now, /dev/tty is implemented as a bit of a hack in open() - it looks for
that absolute path string and if matched, goes to the POSIX service to open
the terminal. Eventually, this should be implemented as a proper device via a
planned user-mode device driver API. However, this is good enough for now to
get Dropbear scp working properly.