Move cache-specific members from page_t into a new structure allocated as needed
page_t had some members that were used only by the page cache. This was
wasteful since these are only needed when pages are actually in use by the
page cache.

This change adds a new page_cache_entry_t structure that is allocated for a
page while it is in use by the cache. There is also some general clean up of
the cache code.

This and the previous change reduce the size of the page database by ~60%.