Add validation to kmalloc() to detect double frees and corruption
Double frees on kmalloc() would happily place the allocation into back into
the slab magazine cache again for reallocation, which would lead to the
buffer being allocated twice.

Add a bunch of validation on the allocation tags in kfree() and krealloc(),
so that we can detect double frees, and other corruption of the tag.