Change IRQ handling to support multiple "domains" instead of a single namespace/controller
In the ARM/FDT world we don't usually have a single interrupt controller or
namespace for IRQ numbers. This is also not really true on modern x86 either,
though while we still only support the legacy PIC we can pretend this is the
case.

This change adds IRQ domains, which are a controller plus a handler table, and
moves all IRQ handling to be managed under a domain.

The device tree now supports setting an IRQ domain on a device, which is used
when registering IRQs for that device. The domain is inherited by all
child devices, which will be used by the DT bus manager to set up domains
according to the FDT.

x86 still just uses a single domain mapped to the PIC. One day we'll add
IOAPIC support...